refactor: extract middleware
